requirements.txt
streamlit==1.32.0
nltk==3.8.1
Pillow==10.1.0
scikit-learn==1.3.2
